+++ assets/js/modules/analytics/datastore/accounts.js
@@ -1,5 +1,5 @@
 /**
- * modules/analytics data store: analytics info.
+ * modules/analytics data store: accounts.
  *
  * Site Kit by Google, Copyright 2020 Google LLC
  *
@@ -32,28 +32,17 @@ import { STORE_NAME } from './index';
 // Actions
 const FETCH_ACCOUNTS_PROPERTIES_PROFILES = 'FETCH_ACCOUNTS_PROPERTIES_PROFILES';
 const FETCH_EXISTING_TAG = 'FETCH_EXISTING_TAG';
-const FETCH_PROPERTIES_PROFILES = 'FETCH_PROPERTIES_PROFILES';
-const FETCH_PROFILES = 'FETCH_PROFILES';
 const RECEIVE_ACCOUNTS_PROPERTIES_PROFILES = 'RECEIVE_ACCOUNTS_PROPERTIES_PROFILES';
 const RECEIVE_ACCOUNTS_PROPERTIES_PROFILES_FAILED = 'RECEIVE_ACCOUNTS_PROPERTIES_PROFILES_FAILED';
 const RECEIVE_EXISTING_TAG = 'RECEIVE_EXISTING_TAG';
 const RECEIVE_EXISTING_TAG_FAILED = 'RECEIVE_EXISTING_TAG_FAILED';
-const RECEIVE_PROPERTIES_PROFILES = 'RECEIVE_PROPERTIES_PROFILES';
-const RECEIVE_PROPERTIES_PROFILES_FAILED = 'RECEIVE_PROPERTIES_PROFILES_FAILED';
-const RECEIVE_PROFILES = 'RECEIVE_PROFILES';
-const RECEIVE_PROFILES_FAILED = 'RECEIVE_PROFILES_FAILED';
 
 export const INITIAL_STATE = {
 	accounts: undefined,
-	// connection: undefined,
-	// isFetchingConnection: false,
 	existingTag: undefined,
 	isFetchingAccountsPropertiesProfiles: false,
-	isFetchingPropertiesProfiles: {},
-	isFetchingProfiles: {},
 	isFetchingExistingTag: false,
-	properties: undefined,
-	profiles: undefined,
+	tagPermissions: {},
 };
 
 export const actions = {
@@ -68,20 +57,6 @@ export const actions = {
 		return {
 			payload: {},
 			type: FETCH_EXISTING_TAG,
-		};
-	},
-
-	*fetchPropertiesProfiles( accountId ) {
-		return {
-			payload: { accountId },
-			type: FETCH_PROPERTIES_PROFILES,
-		};
-	},
-
-	*fetchProfiles( accountId, propertyId ) {
-		return {
-			payload: { accountId, propertyId },
-			type: FETCH_PROFILES,
 		};
 	},
 
@@ -118,48 +93,6 @@ export const actions = {
 			type: RECEIVE_EXISTING_TAG_FAILED,
 		};
 	},
-
-	receivePropertiesProfiles( { accountId, properties, profiles } ) {
-		invariant( accountId, 'accountId is required' );
-		invariant( properties, 'properties is required' );
-
-		return {
-			payload: { accountId, properties, profiles },
-			type: RECEIVE_PROPERTIES_PROFILES,
-		};
-	},
-
-	receivePropertiesProfilesFailed( { accountId, error } ) {
-		invariant( accountId, 'accountId is required' );
-		invariant( error, 'error is required.' );
-
-		return {
-			payload: { accountId, error },
-			type: RECEIVE_PROPERTIES_PROFILES_FAILED,
-		};
-	},
-
-	receiveProfiles( { accountId, propertyId, profiles } ) {
-		invariant( accountId, 'accountId is required' );
-		invariant( propertyId, 'accountId is required' );
-		invariant( profiles, 'profiles is required' );
-
-		return {
-			payload: { accountId, propertyId, profiles },
-			type: RECEIVE_PROFILES,
-		};
-	},
-
-	receiveProfilesFailed( { accountId, error, propertyId } ) {
-		invariant( accountId, 'accountId is required' );
-		invariant( error, 'error is required.' );
-		invariant( propertyId, 'accountId is required' );
-
-		return {
-			payload: { accountId, error, propertyId },
-			type: RECEIVE_PROFILES_FAILED,
-		};
-	},
 };
 
 export const controls = {
@@ -175,12 +108,6 @@ export const controls = {
 		// currently quite nested and difficult to straightforwardly test.
 		return getExistingTag( 'analytics' );
 	},
-	[ FETCH_PROPERTIES_PROFILES ]: ( accountId ) => {
-		return API.get( 'modules', 'analytics', 'properties-profiles', { accountID: accountId } );
-	},
-	[ FETCH_PROFILES ]: ( accountId, profileId ) => {
-		return API.get( 'modules', 'analytics', 'profiles', { accountID: accountId, profileID: profileId } );
-	},
 };
 
 export const reducer = ( state, action ) => {
@@ -189,30 +116,6 @@ export const reducer = ( state, action ) => {
 			return {
 				...state,
 				isFetchingAccountsPropertiesProfiles: true,
-			};
-		}
-
-		case FETCH_PROPERTIES_PROFILES: {
-			const { accountId } = action.payload;
-
-			return {
-				...state,
-				isFetchingPropertiesProfiles: {
-					...state.isFetchingPropertiesProfiles,
-					[ accountId ]: true,
-				},
-			};
-		}
-
-		case FETCH_PROFILES: {
-			const { accountId, propertyId } = action.payload;
-
-			return {
-				...state,
-				isFetchingProfiles: {
-					...state.isFetchingProfiles,
-					[ `${ accountId }::${ propertyId }` ]: true,
-				},
 			};
 		}
 
@@ -265,59 +168,6 @@ export const reducer = ( state, action ) => {
 			};
 		}
 
-		case RECEIVE_PROPERTIES_PROFILES: {
-			const { accountId, properties, profiles } = action.payload;
-
-			return {
-				...state,
-				isFetchingPropertiesProfiles: {
-					...state.isFetchingPropertiesProfiles,
-					[ accountId ]: false,
-				},
-				properties,
-				profiles,
-			};
-		}
-
-		case RECEIVE_PROPERTIES_PROFILES_FAILED: {
-			const { accountId, error } = action.payload;
-
-			return {
-				...state,
-				error,
-				isFetchingPropertiesProfiles: {
-					...state.isFetchingPropertiesProfiles,
-					[ accountId ]: false,
-				},
-			};
-		}
-
-		case RECEIVE_PROFILES: {
-			const { accountId, propertyId, profiles } = action.payload;
-
-			return {
-				...state,
-				isFetchingProfiles: {
-					...state.isFetchingProfiles,
-					[ `${ accountId }::${ propertyId }` ]: false,
-				},
-				profiles,
-			};
-		}
-
-		case RECEIVE_PROFILES_FAILED: {
-			const { accountId, error, propertyId } = action.payload;
-
-			return {
-				...state,
-				error,
-				isFetchingProfiles: {
-					...state.isFetchingProfiles,
-					[ `${ accountId }::${ propertyId }` ]: false,
-				},
-			};
-		}
-
 		default: {
 			return { ...state };
 		}
@@ -340,36 +190,6 @@ export const resolvers = {
 		}
 	},
 
-	*getProperties( accountId ) {
-		try {
-			const response = yield actions.fetchPropertiesProfiles( accountId );
-			const { properties, profiles } = response;
-
-			yield actions.receivePropertiesProfiles( { accountId, properties, profiles } );
-
-			return;
-		} catch ( error ) {
-			// TODO: Implement an error handler store or some kind of centralized
-			// place for error dispatch...
-			return actions.receivePropertiesProfilesFailed( { accountId, error } );
-		}
-	},
-
-	*getProfiles( accountId, propertyId ) {
-		try {
-			const response = yield actions.fetchProfiles( accountId, propertyId );
-			const { profiles } = response;
-
-			yield actions.receiveProfiles( { accountId, propertyId, profiles } );
-
-			return;
-		} catch ( error ) {
-			// TODO: Implement an error handler store or some kind of centralized
-			// place for error dispatch...
-			return actions.receiveProfilesFailed( { accountId, propertyId, error } );
-		}
-	},
-
 	*getExistingTag() {
 		try {
 			const existingTag = yield actions.fetchExistingTag( 'analytics' );
@@ -383,6 +203,19 @@ export const resolvers = {
 			// TODO: Implement an error handler store or some kind of centralized
 			// place for error dispatch...
 			return actions.receiveExistingTagFailed( err );
+		}
+	},
+
+	*getTagPermission( accountId, propertyId ) {
+		try {
+			const existingTag = yield actions.fetchTagPermission( accountId, propertyId );
+			yield actions.receiveTagPermission( existingTag );
+
+			return;
+		} catch ( error ) {
+			// TODO: Implement an error handler store or some kind of centralized
+			// place for error dispatch...
+			return actions.receiveTagPermissionFailed( error );
 		}
 	},
 };
@@ -468,66 +301,34 @@ export const selectors = {
 	},
 
 	/**
-	 * Get all Google Analytics properties this account can access.
+	 * Check permissions for an existing Google Analytics tag.
 	 *
-	 * Returns an array of all analytics properties.
+	 * Get permissions for a tag based on a Google Analytics `accountId` and
+	 * `propertyId`. Useful when an existing tag on the site is found and
+	 * you want to verify that an account + property combination has access to
+	 * said tag.
 	 *
-	 * Returns `undefined` if accounts have not yet loaded.
-	 *
-	 * @since n.e.x.t
-	 *
-	 * @param {Object} state Data store's state.
-	 * @param {string} accountId The Analytics Account ID to fetch properties for.
-	 * @return {Array|undefined} An array of Analytics properties; `undefined` if not loaded.
-	 */
-	getProperties( state, accountId ) {
-		invariant( accountId, 'accountId is required' );
-
-		const { properties } = state;
-
-		if ( properties && properties.length ) {
-			return properties.filter( ( property ) => {
-				return property.accountId === accountId;
-			} );
-		}
-
-		return properties;
-	},
-
-	/**
-	 * Get all Google Analytics profiles this user account+property has available.
-	 *
-	 * Returns an array of all profiles.
-	 *
-	 * Returns `undefined` if accounts have not yet loaded.
+	 * Returns `undefined` if the permission check has not yet loaded.
 	 *
 	 * @since n.e.x.t
-	 *
 	 * @param {Object} state Data store's state.
-	 * @param {string} accountId The Analytics Account ID to fetch profiles for.
-	 * @param {string} propertyId The Analytics Property ID to fetch profiles for.
-	 * @return {Array|undefined} An array of Analytics profiles; `undefined` if not loaded.
+	 * @param {string} accountId The Analytics Account ID to fetch permissions for.
+	 * @param {string} propertyId The Analytics Property ID to check permissions for.
+	 * @param {string} tag The Google Analytics tag identifier to check.
+	 * @return {boolean|undefined} `true` if account + property has permission to access the tag, `false` if not; `undefined` if not loaded.
 	 */
-	getProfiles( state, accountId, propertyId ) {
-		invariant( accountId, 'accountId is required' );
-		invariant( propertyId, 'propertyId is required' );
+	getTagPermission( state, accountId, propertyId, tag ) {
+		const { tagPermissions } = state;
 
-		const { profiles } = state;
-
-		if ( profiles && profiles.length ) {
-			return profiles.filter( ( profile ) => {
-				return (
-					profile.accountId === accountId &&
-					profile.webPropertyId === propertyId
-				);
-			} );
+		if (
+			tagPermissions &&
+			tagPermissions[ accountId ] &&
+			tagPermissions[ accountId ][ propertyId ] !== undefined
+		) {
+			return tagPermissions[ accountId ][ propertyId ].includes( tag );
 		}
 
-		return profiles;
-	},
-
-	getTagPermission() {
-		throw new Error( 'Not yet implemented.' );
+		return undefined;
 	},
 };
